Version 1.1
Alright so now we can actually "play a game" even if it is super crappy. There is now a bunch for you guys to do that is pretty obvious, I also wrote out some "to-do's" in the MatchReal class that everyone could do. Whatever To-do you decide to do, put your name next to that in comments and push that code so multiple people do not work on something at the same time. Do not push anything that is broken ever please, you can always send the code separately if you need help. If you have any questions or need help don't be afraid to ask. Happy coding :)